pyproject.toml
[tool.poetry]
name = "sapron-api"
version = "0.1.0"
description = ""
authors = ["Rafael Paixão <rafaelpx8812@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.2.6"
djangorestframework = "^3.12.4"
psycopg2 = "^2.9.1"
drf-yasg = "^1.20.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
